requirements.txt
streamlit
opencv-python
torch
torchvision
Pillow
numpy
matplotlib
supervision==0.12.0